refactor: 패캐지 strategy -> external로 변경
